Add asset write functions
